requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
Bottleneck==1.3.7
certifi==2024.8.30
cycler==0.11.0
distro==1.9.0
dnspython==2.3.0
exceptiongroup==1.2.2
fonttools==4.43.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
importlib-resources==5.12.0
jiter==0.7.1
kiwisolver==1.4.5
matplotlib==3.9.2
mkl-service==2.4.2
mkl-fft==1.3.11
mkl-random==1.2.8
mysql-connector-python==8.0.31
numexpr==2.8.4
numpy==1.26.4
openai==1.54.4
packaging==24.0
pandas==2.2.2
pillow==11.0.0
ply==3.11
protobuf==3.20.1
pydantic==2.9.2
pydantic-core==2.23.4
pyparsing==3.1.1
PyQt5==5.15.10
PyQt5-sip>=12.13,<13
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.2
sip==6.7.10
six==1.16.0
sniffio==1.3.0
tomli==2.0.1
tornado==6.3.3
tqdm==4.67.0
typing_extensions==4.12.2
tzdata==2023.3
unicodedata2==15.0.0
zipp==3.17.0
